test: add simple response wrapper tests
As the response wrappers will be used in all our responses to the
client, we want to make sure any changes in behaviour do not introduce
breaking changes. The current test suite validate that when called the
wrappers write the correct data on the `http.ResponseWriter` and ensure
that he output is matching the input parameters.