feat(handlers.go): add user creation handler
We will be using an apiConfig struct to keep a reference to our database
as well as an entry point for future business logic. The user creation
will for now only require a name param in its JSON body.

If the JSON decoder it fails to parse the body will return a 400 error.
If the user creation on the database fails, we returns a 500 error instead.
If no error occurs the response is the created users using a custom wrapper to
ensure JSON field are mapped on the expected manner.