refactor: extract auth check into a middleware
This change makes so we can reuse the API key extraction code as a middleware in other authenticated
endpoints. We now add the logic for authentication as a step before the handler  logic is ran.

BREAKING CHANGE: the `handlerGetUser` has had its signature changed and
now needs a param of type `database.User` as input in addition to the
Response writer and request.