refactor(models.go): renamed file
The intent of the code there is to server as API response
representations for the database contents and allow us to control the
format sent to the client. This rename should make that more clear, as
we also have a `models.go` on the `internal/database` package that deals
with the application's data layer.